requirements.txt
PySide6
faster-whisper
ffmpeg-python
tqdm
pydantic
